+++ subsys/net/lib/pandora/wifi.c
@@ -50,7 +50,7 @@ static void stop_dhcpv4_client(struct net_if *iface, void *user_data)
 	net_dhcpv4_stop(iface);
 }
 
-static void wifi_event_handler(struct net_mgmt_event_callback *cb, uint32_t mgmt_event,
+static void wifi_event_handler(struct net_mgmt_event_callback *cb, uint64_t mgmt_event,
 			       struct net_if *iface)
 {
 	switch (mgmt_event) {
